[front] Migrate Opinions.js => Opinions.tsx
